+++ hal/include/HAL/AnalogInput.h
@@ -18,8 +18,8 @@ void HAL_FreeAnalogInputPort(HAL_AnalogInputHandle analog_port_handle);
 HAL_Bool HAL_CheckAnalogModule(int32_t module);
 HAL_Bool HAL_CheckAnalogInputChannel(int32_t pin);
 
-void HAL_SetAnalogSampleRate(float samplesPerSecond, int32_t* status);
-float HAL_GetAnalogSampleRate(int32_t* status);
+void HAL_SetAnalogSampleRate(double samplesPerSecond, int32_t* status);
+double HAL_GetAnalogSampleRate(int32_t* status);
 void HAL_SetAnalogAverageBits(HAL_AnalogInputHandle analog_port_handle,
                               int32_t bits, int32_t* status);
 int32_t HAL_GetAnalogAverageBits(HAL_AnalogInputHandle analog_port_handle,
@@ -33,11 +33,11 @@ int32_t HAL_GetAnalogValue(HAL_AnalogInputHandle analog_port_handle,
 int32_t HAL_GetAnalogAverageValue(HAL_AnalogInputHandle analog_port_handle,
                                   int32_t* status);
 int32_t HAL_GetAnalogVoltsToValue(HAL_AnalogInputHandle analog_port_handle,
-                                  float voltage, int32_t* status);
-float HAL_GetAnalogVoltage(HAL_AnalogInputHandle analog_port_handle,
-                           int32_t* status);
-float HAL_GetAnalogAverageVoltage(HAL_AnalogInputHandle analog_port_handle,
-                                  int32_t* status);
+                                  double voltage, int32_t* status);
+double HAL_GetAnalogVoltage(HAL_AnalogInputHandle analog_port_handle,
+                            int32_t* status);
+double HAL_GetAnalogAverageVoltage(HAL_AnalogInputHandle analog_port_handle,
+                                   int32_t* status);
 int64_t HAL_GetAnalogLSBWeight(HAL_AnalogInputHandle analog_port_handle,
                                int32_t* status);
 int32_t HAL_GetAnalogOffset(HAL_AnalogInputHandle analog_port_handle,
