+++ hal/include/HAL/AnalogOutput.h
@@ -16,8 +16,8 @@ HAL_AnalogOutputHandle HAL_InitializeAnalogOutputPort(
     HAL_PortHandle port_handle, int32_t* status);
 void HAL_FreeAnalogOutputPort(HAL_AnalogOutputHandle analog_output_handle);
 void HAL_SetAnalogOutput(HAL_AnalogOutputHandle analog_output_handle,
-                         float voltage, int32_t* status);
-float HAL_GetAnalogOutput(HAL_AnalogOutputHandle analog_output_handle,
-                          int32_t* status);
+                         double voltage, int32_t* status);
+double HAL_GetAnalogOutput(HAL_AnalogOutputHandle analog_output_handle,
+                           int32_t* status);
 HAL_Bool HAL_CheckAnalogOutputChannel(int32_t pin);
 }
