+++ hal/include/HAL/SPI.h
@@ -39,7 +39,7 @@ void HAL_SetSPIAccumulatorDeadband(int32_t port, int32_t deadband,
 int32_t HAL_GetSPIAccumulatorLastValue(int32_t port, int32_t* status);
 int64_t HAL_GetSPIAccumulatorValue(int32_t port, int32_t* status);
 int64_t HAL_GetSPIAccumulatorCount(int32_t port, int32_t* status);
-float HAL_GetSPIAccumulatorAverage(int32_t port, int32_t* status);
+double HAL_GetSPIAccumulatorAverage(int32_t port, int32_t* status);
 void HAL_GetSPIAccumulatorOutput(int32_t port, int64_t* value, int64_t* count,
                                  int32_t* status);
 }
